dln: Fix bad environment assumptions
This fixes an issue where environment lookups didn't work as expected.

That is, the environment lookup was only (seemingly) right for Windows
environments. It was defaulting to the "Path" environment variable.

This does not fail in conventional POSIX-ish setups, since the fallback
is to look in a bunch of default paths when PATH_ENV is null.

This, however, fails in sandboxed and isolated situations where
dependencies are injected through PATH. This is why it fails on NixOS.

Furthermore, the path separator also defaulted to a non-POSIXly correct
one.

These changes hinge on the Windows conditional, and makes environment
lookups work correctly on non-windows platforms.